+++ django/applications/catmaid/static/js/layers/pixi-tile-layer.js
@@ -192,6 +192,10 @@
     var y = 0;
     var slicePixelPosition = [tileInfo.z];
 
+    // Clamping to zero can be disable in the stack.
+    let clamp = this.stackViewer.primaryStack.clamp;
+    let [minCol, minRow] = clamp ? [0, 0] : [tileInfo.firstCol, tileInfo.firstRow];
+
     // Update tiles.
     for (var i = this._tileOrigR, ti = 0; ti < rows; ++ti, i = (i+1) % rows) {
       var r = tileInfo.firstRow + ti;
@@ -204,8 +208,8 @@
         tile.position.x = x;
         tile.position.y = y;
 
-        if (c >= tileInfo.firstCol && c <= tileInfo.lastCol &&
-            r >= tileInfo.firstRow && r <= tileInfo.lastRow && showTiles) {
+        if (c >= minCol && c <= tileInfo.lastCol &&
+            r >= minRow && r <= tileInfo.lastRow && showTiles) {
           var source = this.tileSource.getTileURL(project, this.stack, slicePixelPosition,
               c, r, tileInfo.zoom);